Final checklist: split test_server_session_gate.py under the 500-line limit
